refactor(tests): extrair TestAuthHandler e ApiFactoryBase para projeto compartilhado
TestAuthHandler e o esqueleto dos ApiFactory eram idênticos entre
Selecao.IntegrationTests e Ingresso.IntegrationTests (diferiam só
pelo namespace e pela string de conexão). Qualquer evolução na
infraestrutura de teste exigia quatro edições sincronizadas.

- Criado projeto tests/Unifesspa.UniPlus.IntegrationTests.Shared com
  TestAuthHandler (autenticação fake para WebApplicationFactory) e
  ApiFactoryBase<TEntryPoint> (factory abstrato que delega overrides
  de configuração para a classe concreta).
- Selecao/IngressoApiFactory viraram classes concretas enxutas que
  só fornecem o connection string do módulo.
- AuthEndpointsTests em Selecao e Ingresso passam a usar TestAuthHandler
  do projeto compartilhado.
- NoWarn=CA1716 no shared: 'Shared' é palavra reservada em VB.NET,
  porém o projeto é test-only e nunca será consumido de VB.

Refs: review do PR #144 (round 2, finding B2 parte 2).
