+++ tests/Unifesspa.UniPlus.Ingresso.IntegrationTests/Infrastructure/IngressoApiFactory.cs
@@ -1,38 +1,15 @@
 namespace Unifesspa.UniPlus.Ingresso.IntegrationTests.Infrastructure;
 
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+using Unifesspa.UniPlus.IntegrationTests.Shared.Hosting;
 
 #pragma warning disable CA1515 // Fixture needs to be visible to xUnit.
-public sealed class IngressoApiFactory : WebApplicationFactory<Program>
+public sealed class IngressoApiFactory : ApiFactoryBase<Program>
 #pragma warning restore CA1515
 {
-    protected override void ConfigureWebHost(IWebHostBuilder builder)
-    {
-        ArgumentNullException.ThrowIfNull(builder);
-
-        builder.UseEnvironment("Development");
-
-        builder.ConfigureAppConfiguration((_, configurationBuilder) =>
-        {
-            configurationBuilder.AddInMemoryCollection(
-            [
-                new KeyValuePair<string, string?>("ConnectionStrings:IngressoDb", "Host=localhost;Port=5432;Database=uniplus_tests;Username=uniplus;Password=uniplus_dev"),
-                new KeyValuePair<string, string?>("Auth:Authority", "http://localhost/test-realm"),
-                new KeyValuePair<string, string?>("Auth:Audience", "uniplus"),
-            ]);
-        });
-
-        builder.ConfigureTestServices(services =>
-        {
-            services.AddAuthentication(TestAuthHandler.SchemeName)
-                .AddScheme<AuthenticationSchemeOptions, TestAuthHandler>(
-                    TestAuthHandler.SchemeName,
-                    _ => { });
-        });
-    }
+    protected override IEnumerable<KeyValuePair<string, string?>> GetConfigurationOverrides() =>
+    [
+        new("ConnectionStrings:IngressoDb", "Host=localhost;Port=5432;Database=uniplus_tests;Username=uniplus;Password=uniplus_dev"),
+        new("Auth:Authority", "http://localhost/test-realm"),
+        new("Auth:Audience", "uniplus"),
+    ];
 }
