+++ tests/Unifesspa.UniPlus.Selecao.IntegrationTests/Unifesspa.UniPlus.Selecao.IntegrationTests.csproj
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <ProjectReference Include="../../src/selecao/Unifesspa.UniPlus.Selecao.API/Unifesspa.UniPlus.Selecao.API.csproj" />
+    <ProjectReference Include="../Unifesspa.UniPlus.IntegrationTests.Shared/Unifesspa.UniPlus.IntegrationTests.Shared.csproj" />
   </ItemGroup>
 
   <ItemGroup>
